tests/fixtures/fixture_user.py
import pytest
from rest_framework.test import APIClient
from rest_framework_simplejwt.tokens import RefreshToken


@pytest.fixture
def user_1(django_user_model):
    return django_user_model.objects.create_user(
        username='user_1', password='123456'
    )


@pytest.fixture
def user_2(django_user_model):
    return django_user_model.objects.create_user(
        username='user_2', password='123456'
    )


@pytest.fixture
def token(user_2):
    refresh = RefreshToken.for_user(user_2)

    return {
        'refresh': str(refresh),
        'access': str(refresh.access_token),
    }


@pytest.fixture
def user_client(token):
    client = APIClient()
    client.credentials(HTTP_AUTHORIZATION=f'Bearer {token["access"]}')
    return client
