_requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.4
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.25.1
exceptiongroup==1.0.4
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==5.0.3
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
isort==5.11.2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.2
packaging==22.0
pep8-naming==0.13.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.6.0
pyrsistent==0.19.2
pytest==7.2.0
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.13
